pyproject.toml
[project]
name = "run-atopile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruamel-yaml>=0.18.10",
]
